feat!: change usize to u32 in signatures for multiple methods
The methods are:
Bounded: fn move_in_current_direction_and_return_new_pos
BoundedMovingObject: fn move_in_current_direction
BoundedMovingObject: fn move_in_direction

The commit also changes the behavior of two internal functions:
Bounded: fn move_in_absolute_direction
Positiond: fn coordinate_in_direction

This is done because u32 makes more sense than usize to represent
magnitude in these cases.

+++ src/bounded.rs
@@ -226,7 +226,7 @@ pub trait Bounded: BoundSeal {
 
     /// Returns true if the coordinate actually moved and false if not, if there is
     /// an attempt to move outside of the border.
-    fn move_in_absolute_direction(&mut self, direction: AbsoluteDirection, magnitude: usize) -> bool
+    fn move_in_absolute_direction(&mut self, direction: AbsoluteDirection, magnitude: u32) -> bool
     where
         Self: Mover,
     {
